% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbisection.R
\name{znormbisection}
\alias{znormbisection}
\title{Z-val Approximate by Using Bisection Method}
\usage{
znormbisection(given_area, max_iteration, li, ui, is_echo = FALSE)
}
\arguments{
\item{given_area}{a given cumulative probability for a given confidence coefficient}

\item{max_iteration}{maximum iteration, the more iteration it have, the more precise it is}

\item{li}{lower limit point estimate for z value (a.k.a. lower bound)}

\item{ui}{upper limit point estimate for z value (a.k.a. upper bound)}

\item{is_echo}{return a vector of estimate point for z value if it's TRUE, by default is_echo is FALSE}
}
\description{
Finding z value for a given cumulative probability of standard normal distribution,
such that probability of Z < z is equal to (1 - alpha) confidence coefficient
using numerical analysis, bisection method.
}
\details{
The probability of z can be calculated by using pnorm(z), when iteration is finish, for
z value, given_area ~ pnorm(z)
}
\examples{
z_approximate <- znormbisection(0.975, 20, 0, 4)
zvect_approximate <- znormbisection(0.975, 20, -4, 4, is_echo= TRUE)
z_approximate
zvect_approximate
}
