% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsimulated_annealing.R
\name{zsimulated_annealing}
\alias{zsimulated_annealing}
\title{Z-val Approximate by Using Simulated Annealing (as supposed)}
\usage{
zsimulated_annealing(given_area, max_iteration, is_echo = FALSE)
}
\arguments{
\item{given_area}{a given cumulative probability for a given confidence coefficient}

\item{max_iteration}{maximum iteration, the more iteration it have, the more precise it get}

\item{is_echo}{return a data frame if it's TRUE, by default is_echo is FALSE}
}
\description{
Finding z value for a given cumulative probability of standard normal distribution,
such that probability of Z < z is equal to 100(1 - alpha) confidence coefficient using
Metaheuristic, simulated annealing (as supposed).
}
\examples{
z_approximate <- zsimulated_annealing(0.975, 20)
zdframe_approximate <- zsimulated_annealing(0.975, 20, is_echo= TRUE)
z_approximate
zdframe_approximate
}
